Moved parser to main, added version arg
