+++ minfluxdbconvert/const.py
@@ -2,7 +2,7 @@
 
 MAJOR_VERSION = 0
 MINOR_VERSION = 0
-PATCH_VERSION = 6
+PATCH_VERSION = 7
 
 __version__ = '{}.{}.{}'.format(MAJOR_VERSION, MINOR_VERSION, PATCH_VERSION)
